+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## tbd
+
+**Breaking changes:**
+- renamed `extendIri` to `extend` and `compactIri` to `compact`
+
+**New features:**
+- Added `registerPrefixes`
+
 ## 0.1.0
 
 First release! :tada: